Update IJ platform, kotlin, compose, remove jewel
