+++ gradle.properties
@@ -3,10 +3,9 @@ jvm-version=21
 org.gradle.jvmargs=-Xmx8g
 # Plugin
 version=0.21
-platform-type=IC
-platform-version=251.25410.59
-since-build=242
-until-build=252.*
+platform-version=2026.1.2
+since-build=251
+until-build=262.*
 change-notes=Move preview after default
 plugin-id=by.overpass.svg-to-compose-intellij
 plugin-name=Svg to Compose
