+++ gradle/libs.versions.toml
@@ -1,22 +1,18 @@
 [versions]
-kotlin = "2.2.10"
-intellij-platform = "2.7.2"
+kotlin = "2.4.0"
+intellij-platform = "2.16.0"
 detekt = "1.23.8"
 junit = "4.13.2"
 kotlin-coroutines = "1.11.0"
 svg-to-compose = "0.11.0"
 mockito-kotlin = "6.3.0"
-jewel = "0.27.0"
 compose-desktop = "1.11.1"
-file-picker = "3.1.0"
 turbine = "1.2.1"
 detekt-compose-rules = "1.4.0"
 lifecycle = "2.9.2"
 
 [libraries]
 svg-to-compose = { group = "com.github.DevSrSouza", name = "svg-to-compose", version.ref = "svg-to-compose" }
-jewel-ide-laf-bridge = { group = "org.jetbrains.jewel", name = "jewel-ide-laf-bridge-241", version.ref = "jewel" }
-compose-multiplatform-file-picker = { group = "com.darkrockstudios", name = "mpfilepicker", version.ref = "file-picker" }
 detekt-compose-rules = { module = "ru.kode:detekt-rules-compose", version.ref = "detekt-compose-rules" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
 mockito-kotlin = { group = "org.mockito.kotlin", name = "mockito-kotlin", version.ref = "mockito-kotlin" }
