+++ src/main/kotlin/by/overpass/svgtocomposeintellij/generator/ui/SvgToComposeDialog.kt
@@ -2,6 +2,7 @@ package by.overpass.svgtocomposeintellij.generator.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -11,6 +12,7 @@ import by.overpass.svgtocomposeintellij.generator.presentation.Finished
 import by.overpass.svgtocomposeintellij.generator.presentation.SvgToComposeState
 import by.overpass.svgtocomposeintellij.generator.presentation.SvgToComposeViewModel
 import by.overpass.svgtocomposeintellij.generator.presentation.isValid
+import by.overpass.svgtocomposeintellij.ui.jewelComposePanelCompat
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogWrapper
 import com.intellij.ui.JBColor
@@ -21,7 +23,6 @@ import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import org.jetbrains.jewel.bridge.JewelComposePanel
 import org.jetbrains.jewel.bridge.toComposeColor
 import org.jetbrains.jewel.foundation.ExperimentalJewelApi
 import org.jetbrains.jewel.foundation.enableNewSwingCompositing
@@ -35,7 +36,7 @@ private const val DEFAULT_DIALOG_CONTENT_WIDTH = 820
 private const val DEFAULT_DIALOG_CONTENT_HEIGHT = 300
 
 class SvgToComposeDialog(
-    project: Project,
+    private val project: Project,
     private val viewModel: SvgToComposeViewModel,
 ) : DialogWrapper(project) {
 
@@ -67,15 +68,17 @@ class SvgToComposeDialog(
     @OptIn(ExperimentalJewelApi::class)
     override fun createCenterPanel(): JComponent {
         enableNewSwingCompositing()
-        return JewelComposePanel {
+        return jewelComposePanelCompat {
             val bgColor by remember(JBColor.PanelBackground.rgb) {
                 mutableStateOf(JBColor.PanelBackground.toComposeColor())
             }
-            SvgToComposePlugin(
-                viewModel = viewModel,
-                modifier = Modifier.fillMaxSize()
-                    .background(bgColor),
-            )
+            CompositionLocalProvider(LocalProject provides project) {
+                SvgToComposePlugin(
+                    viewModel = viewModel,
+                    modifier = Modifier.fillMaxSize()
+                        .background(bgColor),
+                )
+            }
         }.apply {
             preferredSize = JBUI.size(
                 Dimension(
